<CHORE> rename dir host to tenant
