+++ src/main/java/com/instream/tenant/domain/application/domain/entity/ApplicationEntity.java
@@ -2,7 +2,6 @@ package com.instream.tenant.domain.application.domain.entity;
 
 import com.instream.tenant.domain.application.infra.enums.ApplicationType;
 import com.instream.tenant.domain.common.infra.enums.Status;
-import com.instream.tenant.domain.host.domain.entity.TenantEntity;
 import com.instream.tenant.domain.redis.domain.entity.RedisEntity;
 import lombok.*;
 import org.springframework.data.annotation.Id;
