+++ src/main/java/com/instream/tenant/domain/tenant/domain/request/TenantCreateRequest.java
@@ -1,4 +1,4 @@
-package com.instream.tenant.domain.host.domain.request;
+package com.instream.tenant.domain.tenant.domain.request;
 
 public record TenantCreateRequest(
         String account,
